+++ app/mixins/modal/actions.jsx
@@ -1,13 +1,13 @@
-export default class ModalActions {
+export const ModalActionsMixin = {
   show() {
     return true;
-  }
+  },
 
   hide() {
     return true;
-  }
+  },
 
   reset() {
     return true;
   }
-}
+};