+++ app/mixins/modal/store.jsx
@@ -1,0 +1,13 @@
+export const ModalStoreMixin = {
+  showModal: false,
+
+  show() {
+    this.showModal = true;
+  },
+
+  hide() {
+    this.showModal = false;
+  },
+
+  reset() {}
+};