+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
 ## Unreleased
+- Replace abstract actions/stores with their decorator analogs
+([#12](https://github.com/fs/react-base/pull/12))
 - Add lodash to the devDependencies to avoid problem with json-server's resolved lodash version
-([#11](https://github.com/fs/react-base/pull/11))
+([#16](https://github.com/fs/react-base/pull/16))
 - Add sign up functionality
 ([#8](https://github.com/fs/react-base/pull/8))
 - Introduce abstract actions and stores